Adopt article-manager v2 abstract pages; extract ResolvesCommunity
Phase 2 of the reusable-component architecture pass:

- Require jfsullivan/article-manager ^2.0
- New Livewire\Concerns\ResolvesCommunity trait: community_id mount
  property + community() computed + owningModel(), replacing six
  copies of the same boilerplate
- Articles\IndexPage / Articles\ShowPage now extend the new abstract
  Pages\ArticleIndexPage / Pages\ArticleShowPage and declare their
  differences through the layout()/layoutProperties()/view()/
  pageTitle() hooks - no more render() overrides
- The four admin article page subclasses shrink to layout() +
  layoutProperties() on top of ResolvesCommunity
- Delete the broken legacy Articles\EditPage and its
  community.articles.edit route: it referenced members it never had
  (owningModelAlias, model), required an unresolvable mount parameter,
  rendered a nonexistent view, and nothing linked to the route
- Regenerate the phpstan baseline (shrinks 138 -> 124)
