+++ composer.json
@@ -50,7 +50,7 @@
         "brick/money": "^0.8.1",
         "illuminate/contracts": "^11.0||^12.0",
         "jackiedo/timezonelist": "^5.1",
-        "jfsullivan/article-manager": "^1.0",
+        "jfsullivan/article-manager": "^2.0",
         "jfsullivan/common-helpers": "^1.0",
         "jfsullivan/member-manager": "^0.6",
         "jfsullivan/notifications": "^1.0",
